refactor: Divide controller mathod and modify code to suit the manualLotto
